requirements.txt
absl-py==2.1.0
accelerate==1.3.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==25.1.0
axial_positional_embedding==0.3.12
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
cachetools==6.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
clean-fid==0.1.35
click==8.1.8
clip==0.2.0
clip-anytorch==2.6.0
cloudpickle==3.1.1
CoLT5-attention==0.11.1
comm==0.2.2
contourpy==1.3.1
cuda-bindings==12.9.0
cuda-python==12.9.0
cudf-cu12==25.6.0
cugraph-cu12==25.6.0
cuml-cu12==25.6.0
cupy-cuda12x==13.5.0
cuvs-cu12==25.6.1
cycler==0.12.1
dask==2025.5.0
dask-cuda==25.6.0
dask-cudf-cu12==25.6.0
datasets==3.2.0
dctorch==0.1.2
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.32.2
dill==0.3.8
distributed==2025.5.0
distributed-ucxx-cu12==0.44.0
docker-pycreds==0.4.0
einops==0.8.0
executing==2.2.0
fastjsonschema==2.21.1
fastrlock==0.8.3
filelock==3.17.0
fonttools==4.55.8
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.9.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.70.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
hyper-connections==0.2.1
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==8.32.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
json5==0.10.0
jsonmerge==1.9.2
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
k-diffusion==0.1.1.post1
kiwisolver==1.4.8
kornia==0.8.0
kornia_rs==0.1.8
lazy_loader==0.4
libcudf-cu12==25.6.0
libcugraph-cu12==25.6.0
libcuml-cu12==25.6.0
libcuvs-cu12==25.6.1
libkvikio-cu12==25.6.0
libraft-cu12==25.6.0
librmm-cu12==25.6.0
libucx-cu12==1.18.1
libucxx-cu12==0.44.0
lightning==2.5.0.post0
lightning-utilities==0.12.0
llvmlite==0.44.0
local-attention==1.11.2
locket==1.0.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.1
mpmath==1.3.0
msgpack==1.1.1
multidict==6.1.0
multiprocess==0.70.16
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.3.2
notebook_shim==0.2.4
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
nvtx==0.2.12
omegaconf==2.3.0
open_clip_torch==2.30.0
opencv-python==4.11.0.86
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
peft==0.7.0
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
product_key_memory==0.2.11
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pylibcudf-cu12==25.6.0
pylibcugraph-cu12==25.6.0
pylibraft-cu12==25.6.0
pynvjitlink-cu12==0.7.0
pynvml==12.0.0
pyparsing==3.2.1
pysdtw==0.0.5
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
pytorch-lightning==2.5.0.post0
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
raft-dask-cu12==25.6.0
rapids-dask-dependency==25.6.0
rapids-logger==0.1.1
referencing==0.36.2
reformer-pytorch==1.4.4
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rmm-cu12==25.6.0
rpds-py==0.22.3
safetensors==0.5.2
scikit-base==0.12.5
scikit-image==0.25.1
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
Send2Trash==1.8.3
sentencepiece==0.2.0
sentry-sdk==2.20.0
setproctitle==1.3.4
setuptools==75.8.0
six==1.17.0
sktime==0.38.5
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
sympy==1.13.1
tblib==3.1.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2025.1.10
timm==1.0.14
tinycss2==1.4.0
tokenizers==0.21.0
toolz==1.0.0
torch==2.6.0
torch-fidelity==0.3.0
torchdiffeq==0.2.5
torcheval==0.0.7
torchmetrics==1.6.1
torchsde==0.2.6
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trampoline==0.1.2
transformers==4.48.2
treelite==4.4.1
triton==3.2.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
ucx-py-cu12==0.44.0
ucxx-cu12==0.44.0
uri-template==1.3.0
urllib3==2.3.0
wandb==0.19.6
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.13
xformers==0.0.29.post3
xxhash==3.5.0
yarl==1.18.3
zict==3.0.0
zipp==3.21.0
